feat: enhance CodeBlock component with Shiki integration for code highlighting and add utility functions for language handling
